Unit Testing_khadija/DayPeriodModel.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Unit_Testing_khadija
{
    public class DayPeriodModel
    {
        public static string DayPeriod(DateTime dateTime)
        {
            if (dateTime.Hour >= 0 && dateTime.Hour < 6)
            {
                return "Night";
            }
            if (dateTime.Hour >= 6 && dateTime.Hour < 12)
            {
                return "Morning";
            }
            if (dateTime.Hour >= 12 && dateTime.Hour < 18)
            {
                return "Noon";
            }
            return "Evening";
        }

        public static object GetTimeOfDay(DateTime testtime)
        {
            throw new NotImplementedException();
           
        }
    }
}